bugfix + miscellaneous
bugfix = CvUnit::maxCombatStr calls unitCombatDefenseModifier and probably unitCombatAttackModifier with unit combat type -1 (NONE) sometimes. This is apparently caused by selecting a unit with no unit combat type (Construction Ship, Colony Ship) due to the UI's attempts to determine which defender is the best, which it then shows. This gives you a conditionw where the non-combat unit is used as an attacker. Fix: check the unit combat to be passed to make sure it is not -1.

Misc:
CvUnitAI::AI_group - don't let units that can't move due to terrain issues join groups
CvUnit::build + Civ4UnitINfos.xml - make units created from completing improvements spawn with the unit AI specified as their default in the unit info XML which is now set it to carrier sea (starbases) or spy sea (sensor stations).