Many AI adjustments
More adjustments for actually getting the AI to produce a few units with the unit AI for carriers, missile carriers, and things to be carried.

Adjustments to get it to defend its cities a bit better.

Add value to civics for the iPlanetPopCapIncrease property. 

Add value for building selection for PlanetYieldChanges, TraitPlanetYieldChanges, and iPlanetPopCapIncrease.

Add value for Tech selection for buildings at tech with things in above comment.

Fix 2 bugs in CvPlayerAI::AI_bestTech - 
a) it was calculating but not applying a value for buildings that reduce maintenance
b) when in financial trouble it was adding in extra value for maintenance change, but good MaintenanceModifier values are negative (changed to subtract instead)

More carrier adjustment - that had been sitting around in cities waiting for units to group with them, but those units it was checking for had sea type unit AIs that are not used... Also don't waste time looping over map looking for coastal cities to blockade.

More missile loading adjustments (add even more UNITAI types to try to load missiles onto).

Increase range to check for "closeness" factor a bit.

Slight reduction in the number of UNITAI_ATTACK units that it is willing to have wandering around exploring (which was pushing the number of exploration units up 1 over the number CvPlayer::AI_neededExplorers calculated it needed)/

Adjust "chipotle" mode plot hover with Ctrl pressed to show cargo.