Wacky unit AI types bug fix + Spy unit AI tweak.
Fixes a bug introduced by the seemingly simple change of a ">" to a ">=" in CvCityAI::AI_bestUnitAI in Rev 41. While I was at it, it made the function more efficient when there is no suitable unit for the specified unit AI type.

Also tweaked the CvPlayerAI::AI_unitValue function for UNITAI_SPY so it should be less likely to pick units like Colony Ships for it.