Another AstroTech tweak
AstroTech now gets the free promotion on all missiles too, including free ones
generated by starbases if the option to disable their production is not used.
There is no way for anyone else to get this (the promotion is not normally available
to missiles), so AstroTech's missiles are now 10% stronger than those of anybody else.