Healing Promotion Rearrangement
Adjusting the healing promotions was irritating due to the promotion upgrade tree in the Sevopedia having some issues...
When the new PROMOTION_REPAIR_FACILITIES_II was placed in the file directly after PROMOTION_REPAIR_FACILITIES the promotion tree generator python would fail. Apparently it tries to remove key 21 from a list twice. The really weird thing about that is that the new promotion was not number 21, it was number 32. Promotion 21 is the Advanced Bridge.
Putting the new promotion at the end worked. Moving both the new one and it's first level version (PROMOTION_REPAIR_FACILITIES) to a later position in the file also works, so they are currently located after the Ace promotion even though the promotion list would look better with them immediately after the 3 repair staff promotions (which is where the Repair Facilities promotion originally was).
Note that this problem exists whether the Repair Facilities requires one of either Upgraded Weapons or Upgraded Armor or if it requires just Upgraded Repair Staff, which it now does, even though requiring Upgraded Repair Staff moves it out of the main promotion tree where Advanced Bridge is located. It is very strange (and possibly flat out stupid).