Sevopedia bGraphicalOnly tweaks + related Terrain bGraphicalOnly settings
With the Python adjustment, the Sevopedia should now not list anything that has bGraphicalOnly set to 1. All of the terrain types except "Space" are now set to have bGraphicalOnly = 1. The various improvements that should not have shown up were already set to have this (so nothing needed to be done). The minor civilization type and "teh alienz" are still showing up - will be fixed later.