Bugfix for troop transport in unit XML, also a Troop Tranport related AI tweak.
Bug was: Forge UU for invasion ship does not have flag enabling it to capture star systems if the Troop Transport game option is active (so the Forge have no way to capture a star system).

Also added some adjustments to the value for the attack city unit AI type for the no capture setting and the troop transport game option and the troop transport setting.