Oops, forgot nebula damage
Since it is now possible to enter plots with Nebula, it is set to do 10% damage, -5% with the Advanced Shields promotion. The Hardened Shields promotion does not reduce the damage since it is required to enter them at all so any reduction is already included in the base damage for the feature. Might want to change this if a combat ship that can enter without the promotion is ever added.