Added option for star system numbers for both galaxy maps.
It shows 3 settings: Random, Normal, and Dense

Dense gives +20% star systems, Normal is what it was.

I also adjusted the SolarSystemsPerPlot value in the WormholeSpiralGalaxy.py script to match the one in SipralGalaxy.py - it was .0042 and is now .0043. On a standard size map that should result in about 2 more star systems. The BlackHolesPerPlot and SupernovasPerPlot values are still slightly lower, but the wormholes should mroe or less make up for it in overall number of thigns on the map.