Adjustments for starving systems.
Changes have a comment including "post v1.81".

This includes a bug fix for an issue in updatePlotYield where after capturing a star system that has a lunar base outside the current culture range that planet would contribute a negative amount of food. The calculation loops over all planets and those outside the culture range add nothing because the have 0 population. Unfortunately the way the lunar base worked it made such a planet effectively have a population of -1 for the food calculation, so it subtracted food (the case I saw was for a planet with a food yield of 4, so it was contributing -4 which). It now prevents the calculated population from going below 0.