A bugfix for widget help and improvement for starabase culture radius.
Both of these were done due to the B5 mod, which is still being worked on a lot.

The bug in the widget help did not appear in FFP since we never had enough actions for a unit to hit it. If there are enough actions, like in B5 now, then the iData1 values overlap and it tries to get planet data for things that are not planet related. Fixed by checking the widget type.

The culture change is to actually make use of the culture radius specified in the UnitInfo instead of always just using a value of 2. The Omega Starbase was updated to give it a culture range of 3.